+++ lib/pronto/config.rb
@@ -4,7 +4,7 @@ module Pronto
       @config_hash = config_hash
     end
 
-    %w[github gitlab bitbucket].each do |service|
+    %w[github].each do |service|
       ConfigFile::EMPTY[service].each do |key, _|
         name = "#{service}_#{key}"
         define_method(name) { ENV["PRONTO_#{name.upcase}"] || @config_hash[service][key] }
@@ -52,10 +52,6 @@ module Pronto
 
     def github_hostname
       URI.parse(github_web_endpoint).host
-    end
-
-    def bitbucket_hostname
-      URI.parse(bitbucket_web_endpoint).host
     end
 
     def warnings_per_review